+++ dustbuster/Assets/Koyama/Scenes/GameClear.unity
@@ -723,8 +723,8 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 770680894}
-  - component: {fileID: 770680898}
   - component: {fileID: 770680897}
+  - component: {fileID: 770680898}
   - component: {fileID: 770680896}
   - component: {fileID: 770680895}
   m_Layer: 5
@@ -799,7 +799,7 @@ MonoBehaviour:
     m_PersistentCalls:
       m_Calls:
       - m_Target: {fileID: 770680898}
-        m_TargetAssemblyTypeName: S_SwitchController, Assembly-CSharp
+        m_TargetAssemblyTypeName: SceneController, Assembly-CSharp
         m_MethodName: SwitchScene
         m_Mode: 1
         m_Arguments:
@@ -857,7 +857,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 770680893}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: ee5a6ff33d26dcb4f8febe6fe41f60df, type: 3}
+  m_Script: {fileID: 11500000, guid: 1768aac84e75a1447b6364737f19c9f4, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   sceneName: Menu
