--- dustbuster/Assets/Koyama/Scripts/Object/TimeActiveObj.cs
@@ -1,26 +1,0 @@
-//using UnityEngine;
-//using System.Collections;
-
-//public class TimeActiveObj : MonoBehaviour
-//{
-
-//    public float life_time = 1.5f;
-//    float time = 0f;
-
-//    // Use this for initialization
-//    void Start()
-//    {
-//        time = 0;
-//    }
-
-//    // Update is called once per frame
-//    void Update()
-//    {
-//        time += Time.deltaTime;
-//        print(time);
-//        if (time > life_time)
-//        {
-//            GameObject.SetActive(true);
-//        }
-//    }
-//}
