--- dustbuster/Assets/Koyama/Scripts/Object/TimeActiveObj.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a97f682eefa43d948aea6330c77dd325
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
