--- dustbuster/Assets/Koyama/Scripts/PopupScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e4a04b56cd0ec164b8c3dd9a49f5ffbc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
