+++ dustbuster/Assets/Koyama/Scripts/SceneChange.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f42971ed7a14cc14fbac63485283bc4f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
