+++ dustbuster/Assets/Koyama/Scripts/SceneChange/S_SwitchController.cs
