--- dustbuster/Assets/Koyama/Scripts/SceneController.cs
@@ -1,17 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-using UnityEngine.UI;　　　//ボタンを使用するのでUI
-using UnityEngine.SceneManagement;//SceneManagerを使用ためSceneManagementを追加
-
-public class SceneController : MonoBehaviour
-{
-    public string sceneName;
-
-    // ボタンをクリックするとScene『B』に移動
-    public void ButtonClick()
-    {
-        SceneManager.LoadScene(sceneName);
-    }
-}
