--- dustbuster/Assets/Koyama/Scripts/StageScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: adbb485229c0a4e46a678d4f1247f7f3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
